pyproject.toml
[project]
name = "customer-support-chatbot"
version = "0.1.0"
description = "Customer support chatbot using OpenAI API"
authors = [
    {name = "nadja nadjacarabulea@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "openai (>=1.97.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "flask-session[redis] (>=0.8.0,<0.9.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
